Move theme folder to components/theme-legacy as a reminder to clean these up
