+++ Source/SelfService/Web/admin/customer/create.tsx
@@ -8,7 +8,7 @@ import { useSnackbar } from 'notistack';
 
 import { Box, Grid, Step, Stepper, StepContent, StepLabel, Typography } from '@mui/material';
 
-import { TextField as ThemedTextField } from '../../theme/textField';
+import { TextField as ThemedTextField } from '../../components/theme-legacy/textField';
 import { Button } from '@dolittle/design-system';
 
 import { createCustomer, HttpCustomerRequest } from '../../api/solutions/customer';
