+++ Source/SelfService/Web/solutions/backup/viewCard.tsx
@@ -12,7 +12,7 @@ import {
 
 import { mergeStyles } from '@fluentui/react/lib/Styling';
 import { HttpResponseApplication } from '../../api/solutions/application';
-import { cardStyles, commandTileClass, buttonStyles } from '../../theme/viewCard';
+import { cardStyles, commandTileClass, buttonStyles } from '../../components/theme-legacy/viewCard';
 import { getLatestBackupLinkByApplication } from '../../api/solutions/backups';
 import { useGlobalContext } from '../../context/globalContext';
 import { Typography } from '@mui/material';
