+++ Source/SelfService/Web/solutions/documentation/cicd/setupAzurePipelines.tsx
@@ -6,7 +6,7 @@ import { useSnackbar } from 'notistack';
 import Input from '@mui/material/Input';
 
 
-import { ButtonText } from '../../../theme/buttonText';
+import { ButtonText } from '../../../components/theme-legacy/buttonText';
 import { getAzureDevopsKubernetesServiceAccount, getContainerRegistry } from '../../../api/solutions/cicd';
 import { Info } from '../../stores/documentationInfo';
 import { Typography } from '@mui/material';
